add button comp.
